# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记账 API。部分业务错误使用 HTTP 200，通过 status/msg 判断结果。",
        "title": "Jiezhang API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/header": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "首页头部数据（收支趋势、预算等）",
                "operationId": "HomeHandler_Header",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.HomeHeaderResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/index": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "首页账单列表，Query: `range`（today/yesterday/week/month/year）",
                "operationId": "HomeHandler_Index",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "range",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.StatementListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/message": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "消息列表",
                "operationId": "MessagesHandler_List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.MessageListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/message/{id}": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "消息详情",
                "operationId": "MessagesHandler_Show",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.MessageDetailItem"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "用户设置，返回 `{user, version}`，`user.avatar_url` 已按 Rails avatar_path 逻辑处理",
                "operationId": "HomeHandler_GetSettings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.HomeSettingsResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.HomeHeaderMessage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "sub_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.HomeHeaderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/types.HomeHeaderMessage"
                },
                "month_budget": {
                    "type": "string"
                },
                "month_expend": {
                    "type": "string"
                },
                "today_expend": {
                    "type": "string"
                },
                "trends": {
                    "$ref": "#/definitions/types.HomeHeaderTrends"
                },
                "use_pencentage": {
                    "type": "integer"
                }
            }
        },
        "types.HomeHeaderTrendItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "ratio": {
                    "type": "number"
                },
                "trend": {
                    "type": "string"
                }
            }
        },
        "types.HomeHeaderTrends": {
            "type": "object",
            "properties": {
                "day": {
                    "$ref": "#/definitions/types.HomeHeaderTrendItem"
                },
                "month": {
                    "$ref": "#/definitions/types.HomeHeaderTrendItem"
                },
                "week": {
                    "$ref": "#/definitions/types.HomeHeaderTrendItem"
                }
            }
        },
        "types.HomeSettingsAccountBook": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.HomeSettingsResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/types.HomeSettingsUser"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "types.HomeSettingsUser": {
            "type": "object",
            "properties": {
                "account_book": {
                    "$ref": "#/definitions/types.HomeSettingsAccountBook"
                },
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "persist": {
                    "type": "integer"
                },
                "remind": {
                    "type": "boolean"
                },
                "show_diamond": {
                    "type": "boolean"
                },
                "theme": {
                    "$ref": "#/definitions/types.Theme"
                },
                "theme_id": {
                    "type": "integer"
                },
                "themes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Theme"
                    }
                },
                "uid": {
                    "type": "integer"
                }
            }
        },
...
        "types.MessageDetailItem": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "msg_type": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.MessageListItem": {
            "type": "object",
            "properties": {
                "already_read": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "msg_type": {
                    "type": "string"
                },
                "page_url": {
                    "type": "string"
                },
                "sub_title": {
                    "type": "string"
                },
                "target_type": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "types.StatementListItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "asset": {
                    "type": "string"
                },
                "asset_id": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_pic": {
                    "type": "boolean"
                },
                "icon_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "money": {
                    "type": "string"
                },
                "month_day": {
                    "type": "string"
                },
                "mood": {
                    "type": "string"
                },
                "payee": {
                    "$ref": "#/definitions/types.StatementPayee"
                },
                "province": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "target_object": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "timeStr": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "week": {
                    "type": "string"
                }
            }
        },
...
        "types.StatementPayee": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "types.Theme": {
            "type": "object",
            "properties": {
                "class_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppID": {
            "type": "apiKey",
            "name": "X-WX-APP-ID",
            "in": "header"
        },
        "SessionKey": {
            "type": "apiKey",
            "name": "X-WX-Skey",
            "in": "header"
        }
    }
}